\name{Covariates}
\alias{Covariates}
\docType{data}
\title{
An example of Covariate objects
}
\description{
A (100 x 2) matrix. The first column is the intercept and filled with 1s for all samples. The second column is a randomly generated covariate consisting of two classes, 0 and 1.
}
\examples{
data(sampledata)
dim(Covariates)#100 rows and 2 columns
}
\keyword{datasets}
