\name{Geno}
\alias{Geno}
\docType{data}
\title{
An example of marker genotype objects
}
\description{
An example of a marker genotype object consisting of 1000 markers and 100 samples. The genotypes are coded as 0 (AA), 1 (AB), and 2 (BB). The genotypes were randomly generated as described in the details.
}
\details{
Geno was generated by\cr\cr
N<-100\cr
P<-1000\cr
Geno<-matrix(sample(c(0,1,2),N*P,replace=T,\cr
prob=c(0.49,0.42,0.09)),nc=P)\cr
#MAF is 0.3
}
\seealso{
  hyperpara
}
\examples{
data(sampledata)
dim(Geno)#100 samples and 1000 markers
unique(Geno[1:(100*1000)])#0,1,2
}
\keyword{datasets}
