\name{X}
\alias{X}
\docType{data}
\title{
An example of SNP genotypes (explanatory variables)
}
\description{
An example of SNP genotypes consisting of 1000 SNPs and 500 samples. The genotypes are coded as 0 (AA), 1 (AB), and 2 (BB). The genotypes were randomly generated as described in the details. The first, 101th, ..., and 901th SNP have additive effects. The first, 201th, ..., and 801th SNPs have also dominance effects.
}
\details{
X was generated by\cr\cr
N <- 500\cr
P <- 1000\cr
X <- matrix(sample(c(0, 1, 2), N * P, replace = T,\cr
prob = c(0.49, 0.42, 0.09)), nc = P)\cr
}
\seealso{
  Y, Z
}
\examples{
data(sampledata)
dim(X) #500 samples and 1000 SNPs
unique(X[1:(500*1000)]) #0,1,2
}
\keyword{datasets}
